requirements.txt
streamlit==1.32.0
pandas==2.2.0
plotly==5.18.0
pymysql==1.1.0
bcrypt==4.1.2
google-generativeai==0.3.2
python-dotenv==1.0.0